+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - **Project structure:** Enforced project structure.
 - **Module encapsulation:** Enforced module encapsulation.
 - **Path**: Updated path rules package to flat config.
+- **Wisemen:** Made custom rules package. Enabled rule `explicit-function-return-type-with-regex` and made it so composables don't require a return type.
 
 ## [1.5.1] (2025-06-05) 
 - **Tailwindcss:** enforce-consistent-variable-syntax rule ignores classes with custom-.